Getting summary statistics for data with binary outcome

Getting summary statistics for data with binary outcome

Usage

get.summstat.binary(Y, X, B, A, method = "all")

Arguments

Y

a binary vector of outcomes

X

a binary vector of exposures

B

a matrix of binary covariates, with each column representing a covariate

A

a matrix of categorical covariates, with each column representing a covariate

method

specify the type of summary statistics abstracted corresponding to the data generation method, with options including 'all', 1, 2, 3. 'all' will return all summary statistics can be used for generate data with three methods, while 1, 2, 3 corresponding to the multivariate normal thresholding for A and B, the multivariate normal thresholding only for B, and the chain of regression methods, respectively. See more details in the manuscript.

Value

a list of summary statistics